Complet + Log + State

+++ EasySaveConsole/EasySaveConsole/Models/Log.cs
@@ -12,7 +12,7 @@ namespace EasySaveConsole.Models
         public string Name { get; set; }
         public string FileSource { get; set; }
         public string FileTarget { get; set; }
-        public int FileSize { get; set; }
-        public float FileTransferTime { get; set; }
+        public long FileSize { get; set; }
+        public long FileTransferTime { get; set; }
     }
 }
